Move common packages to a separate folder
