+++ src/__tests__/contexts/FormWrapper.test.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { render, fireEvent, screen } from "@testing-library/react";
-import * as formWrapper from "../../contexts/FormWrapper";
+import * as formWrapper from "../../contexts/FormWrapper.example";
 import * as reactRouter from "react-router-dom";
-import useOfflineStorage from "../../hooks/useOfflineStorage";
+import useOfflineStorage from "../../hooks/useOfflineStorage.example";
 
 // Mocking react-router-dom hooks
 vi.mock("react-router-dom", async () => ({
