+++ src/components/HeaderNav.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Title, Tag } from "@trussworks/react-uswds";
-import { Header, Navigation } from "../react-radfish";
+import { Header, Navigation } from "../packages/react-components";
 import { version } from "../../package.json";
 import Logo from "../assets/noaa-logo-circle.svg";
