+++ src/contexts/FormWrapper.example.jsx
@@ -7,10 +7,10 @@
 import React, { createContext, useState, useCallback, useEffect } from "react";
 import { useNavigate, useSearchParams, useParams } from "react-router-dom";
 import { Alert } from "@trussworks/react-uswds";
-import { Form, Button } from "../react-radfish";
+import { Form, Button } from "../packages/react-components";
 import { FORM_CONFIG } from "../config/form";
-import RadfishAPIService from "../services/APIService";
-import useOfflineStorage from "../hooks/useOfflineStorage";
+import RadfishAPIService from "../packages/services/APIService";
+import useOfflineStorage from "../hooks/useOfflineStorage.example";
 import { COMMON_CONFIG } from "../config/common";
 
 const FormContext = createContext();
