+++ src/pages/ComplexForm.example.jsx
@@ -1,9 +1,16 @@
 import React from "react";
 import { Alert, Checkbox, FormGroup } from "@trussworks/react-uswds";
-import { TextInput, Radio, Select, Button, Label, ErrorMessage } from "../react-radfish";
-import { useFormState } from "../contexts/FormWrapper";
+import {
+  TextInput,
+  Radio,
+  Select,
+  Button,
+  Label,
+  ErrorMessage,
+} from "../packages/react-components";
+import { useFormState } from "../contexts/FormWrapper.example";
 import { fullNameValidators } from "../utilities";
-import useOfflineStorage from "../hooks/useOfflineStorage";
+import useOfflineStorage from "../hooks/useOfflineStorage.example";
 import { CONSTANTS } from "../config/form";
 import "../styles/theme.css";
