+++ src/pages/Table.example.jsx
@@ -4,10 +4,9 @@
  */
 
 import { useEffect, useState } from "react";
-import { useTableState } from "../contexts/TableWrapper";
+import { useTableState } from "../contexts/TableWrapper.example";
 import { MSW_ENDPOINT } from "../mocks/handlers";
-import RadfishAPIService from "../services/APIService";
-import { Toast, TOAST_CONFIG } from "../react-radfish";
+import RadfishAPIService from "../packages/services/APIService";
 import {
   Table,
   TableBody,
@@ -21,10 +20,12 @@ import {
   TablePaginationPageCount,
   TablePaginationGoToPage,
   TablePaginationSelectRowCount,
-} from "../react-radfish";
+  Toast,
+  ToastStatus,
+} from "../packages/react-components";
 import { useNavigate } from "react-router-dom";
 const TOAST_LIFESPAN = 3000;
-import useOfflineStorage from "../hooks/useOfflineStorage";
+import useOfflineStorage from "../hooks/useOfflineStorage.example";
 import { Alert, Grid } from "@trussworks/react-uswds";
 import { COMMON_CONFIG } from "../config/common";
 import { GridContainer } from "@trussworks/react-uswds";
@@ -123,10 +124,10 @@ const SimpleTable = () => {
       //delete submitted drafts from local storage
       const idsFromApiResponse = data.map((item) => item.id);
       deleteOfflineData("formData", idsFromApiResponse);
-      const { status, message } = TOAST_CONFIG.SUCCESS;
+      const { status, message } = ToastStatus.SUCCESS;
       setToast({ status, message });
     } catch (error) {
-      const { status, message } = TOAST_CONFIG.ERROR;
+      const { status, message } = ToastStatus.ERROR;
       setToast({ status, message });
     } finally {
       setTimeout(() => {
